バースト銃の実装中。 現状 バースト中に射撃をやめると射撃が止まるので一度売ったら決められた数の弾を絶対に撃つようにする

+++ class/GameObjects/Weapon/Melee/IMelee.cpp
@@ -58,7 +58,7 @@ void IMelee::DebugGui() {
 			}
 			// 弾
 			if (ImGui::TreeNode("Bullet")) {
-				ImGui::DragFloat("Num", &data_.bulletNum);
+				ImGui::DragInt("Num", &data_.bulletNum);
 				ImGui::DragFloat("Speed", &data_.bulletSpeed);
 				ImGui::TreePop();
 			}
